[CCUBE-1472][JY] Shift shared modal styles to modal.styles. Use sub-directory imports. Rename event handlers.
