+++ src/components/common/modal/bulk-edit-modal.styles.tsx
@@ -3,6 +3,6 @@ import styled from "styled-components";
 export const Wrapper = styled.div`
     display: flex;
     flex-direction: column;
-    gap: 16px;
+    gap: 1rem;
     text-align: left;
 `;
