+++ src/components/common/modal/discard-changes-modal.tsx
@@ -1,12 +1,9 @@
 import { ErrorDisplay } from "@lifesg/react-design-system/error-display";
 import { IDiscardChangesModalProps } from "src/context-providers";
 import { useModal } from "../../../context-providers/display/modal-hook";
-import {
-    ActionButton,
-    ButtonWrapper,
-    ModalDisplayTitle,
-} from "./discard-changes-modal.styles";
+import { ModalDisplayTitle } from "./discard-changes-modal.styles";
 import { GenericModal } from "./generic-modal";
+import { ActionButton, ButtonWrapper } from "./modals.styles";
 
 interface IProps {
     modal: IDiscardChangesModalProps;
