+++ src/components/common/modal/modals.styles.tsx
@@ -1,3 +1,4 @@
+import { Button } from "@lifesg/react-design-system/button";
 import { Modal } from "@lifesg/react-design-system/modal";
 import styled from "styled-components";
 
@@ -8,4 +9,15 @@ export const ScrollableModal = styled(Modal)`
         overflow-y: auto;
     }
     cursor: pointer;
+`;
+
+export const ButtonWrapper = styled.div`
+    display: flex;
+    justify-content: center;
+    margin-top: 2rem;
+    gap: 2rem;
+`;
+
+export const ActionButton = styled(Button.Default)`
+    width: 100%;
 `;
