+++ src/components/element-editor/basic-details/dropdown-preselected-value.tsx
@@ -1,6 +1,6 @@
 import { Form } from "@lifesg/react-design-system/form";
 import { Text } from "@lifesg/react-design-system/text";
-import { isEmpty } from "lodash";
+import isEmpty from "lodash/isEmpty";
 import { useEffect } from "react";
 import { Controller, useFormContext } from "react-hook-form";
 import { IDropdownItemAttributes } from "src/context-providers/builder/";
