+++ src/schemas/bulk-edit-helper.ts
@@ -1,4 +1,4 @@
-import { isEmpty } from "lodash";
+import isEmpty from "lodash/isEmpty";
 
 export const getNonEmptyLines = (value: string): string[] => {
     return value.split("\n").filter((line) => !isEmpty(line));
